This is a version modify INT RA2 of hall input signal,version -05 data-20190114

+++ nHIOS_PIC16F1828.X/Delay.c
@@ -23,7 +23,7 @@
  * input regerence :
  *
  *************************************************************************/
-    void delay_10ms(uint t)
+    void delay_10ms(ulong t)
     {
         uint i= 630 *t ; // 8MHZ
         while(i --)
